+++ client/src/apiTypes.ts
@@ -12,6 +12,7 @@ import type { TrackerId } from "./game/systems/trackers/models";
 
 export type ApiShape = ApiAssetRectShape | ApiRectShape | ApiCircleShape | ApiCircularTokenShape | ApiPolygonShape | ApiTextShape | ApiLineShape | ApiToggleCompositeShape
 export type ApiDataBlock = ApiRoomDataBlock | ApiShapeDataBlock | ApiUserDataBlock
+export type ApiShapeCustomData = ApiShapeCustomDataText | ApiShapeCustomDataNumber | ApiShapeCustomDataBoolean | ApiShapeCustomDataDiceExpression
 
 /* eslint-disable */
 /**
@@ -174,6 +175,7 @@ export interface ApiCoreShape {
   ignore_zoom_size: boolean;
   is_door: boolean;
   is_teleport_zone: boolean;
+  custom_data: ApiShapeCustomData[];
   owners: ApiShapeOwner[];
   trackers: ApiTracker[];
   auras: ApiAura[];
@@ -371,6 +373,32 @@ export interface ApiOptionalUserOptions {
 export interface ApiRoomDataBlock extends ApiCoreDataBlock {
   category: "room";
   data: string;
+}
+export interface ApiShapeCustomDataBoolean extends ApiShapeCustomDataCore {
+  kind: "boolean";
+  value: boolean;
+}
+export interface ApiShapeCustomDataCore extends ApiShapeCustomDataIdentifier {
+  kind: string;
+  description?: string;
+}
+export interface ApiShapeCustomDataDiceExpression extends ApiShapeCustomDataCore {
+  kind: "dice-expression";
+  value: string;
+}
+export interface ApiShapeCustomDataIdentifier {
+  shapeId: GlobalId;
+  source: string;
+  prefix: string;
+  name: string;
+}
+export interface ApiShapeCustomDataNumber extends ApiShapeCustomDataCore {
+  kind: "number";
+  value: number;
+}
+export interface ApiShapeCustomDataText extends ApiShapeCustomDataCore {
+  kind: "text";
+  value: string;
 }
 export interface ApiShapeDataBlock extends ApiCoreDataBlock {
   category: "shape";