+++ CHANGELOG.md
@@ -10,6 +10,14 @@ tech changes will usually be stripped from release notes for the public
 
 ## Unreleased
 
+### Added
+
+- Custom shape data
+    - A new tab in the Shape settings UI
+    - Can be used to freely add custom data and create a rudimentary character sheet
+    - Currently supported data types are text, number, toggle and dice expressions
+    - Dice expressions can reference other data entries and open the dice tool with the evaluated expression
+
 ### Changed
 
 - Initiative